# doc/openapi.json
{"openapi":"3.1.0","info":{"title":"SmartStore","version":"0.1.0"},"paths":{"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/token":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_api_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cart/":{"get":{"tags":["cart"],"summary":"List Cart","operationId":"list_cart_api_cart__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CartItemOut"},"type":"array","title":"Response List Cart Api Cart  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["cart"],"summary":"Add Item","operationId":"add_item_api_cart__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartItemCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CartItemOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/cart/{cart_item_id}":{"delete":{"tags":["cart"],"summary":"Delete Item","operationId":"delete_item_api_cart__cart_item_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"cart_item_id","in":"path","required":true,"schema":{"type":"integer","title":"Cart Item Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cart/clear":{"delete":{"tags":["cart"],"summary":"Clear User Cart","operationId":"clear_user_cart_api_cart_clear_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/chat":{"post":{"summary":"REST 对话接口","description":"请求体： {\"text\": \".\"}\n返回：   {\"reply\": \".\", \"actions\":[.]}","operationId":"chat_endpoint_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/orders/":{"get":{"tags":["orders"],"summary":"List Orders","operationId":"list_orders_api_orders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderOut"},"type":"array","title":"Response List Orders Api Orders  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["orders"],"summary":"Place Order","operationId":"place_order_api_orders__post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/orders/all":{"get":{"tags":["orders"],"summary":"List All Orders","operationId":"list_all_orders_api_orders_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderOut"},"type":"array","title":"Response List All Orders Api Orders All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/products/":{"post":{"tags":["products"],"summary":"Create","operationId":"create_api_products__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["products"],"summary":"List Products","operationId":"list_products_api_products__get","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string","description":"搜索关键词","default":"","title":"Q"},"description":"搜索关键词"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductOut"},"title":"Response List Products Api Products  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/products/{sku}":{"get":{"tags":["products"],"summary":"Get Sku","operationId":"get_sku_api_products__sku__get","parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string","title":"Sku"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["products"],"summary":"Update Sku","operationId":"update_sku_api_products__sku__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string","title":"Sku"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["products"],"summary":"Delete Sku","operationId":"delete_sku_api_products__sku__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string","title":"Sku"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/":{"get":{"tags":["users"],"summary":"List Users","operationId":"list_users_api_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserOut"},"type":"array","title":"Response List Users Api Users  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/users/{username}":{"delete":{"tags":["users"],"summary":"Delete User","operationId":"delete_user_api_users__username__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_for_access_token_api_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_auth_token_post"},"CartItemCreate":{"properties":{"sku":{"type":"string","title":"Sku"},"quantity":{"type":"integer","minimum":1.0,"title":"Quantity"}},"type":"object","required":["sku","quantity"],"title":"CartItemCreate"},"CartItemOut":{"properties":{"sku":{"type":"string","title":"Sku"},"quantity":{"type":"integer","minimum":1.0,"title":"Quantity"},"cart_item_id":{"type":"integer","title":"Cart Item Id"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"product":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Product"}},"type":"object","required":["sku","quantity","cart_item_id","added_at","product"],"title":"CartItemOut"},"ChatRequest":{"properties":{"text":{"type":"string","title":"Text"}},"type":"object","required":["text"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OrderItemOut":{"properties":{"sku":{"type":"string","title":"Sku"},"quantity":{"type":"integer","title":"Quantity"},"unit_price":{"type":"integer","title":"Unit Price"}},"type":"object","required":["sku","quantity","unit_price"],"title":"OrderItemOut"},"OrderOut":{"properties":{"order_id":{"type":"integer","title":"Order Id"},"total_cents":{"type":"integer","title":"Total Cents"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"items":{"items":{"$ref":"#/components/schemas/OrderItemOut"},"type":"array","title":"Items"}},"type":"object","required":["order_id","total_cents","status","created_at","items"],"title":"OrderOut"},"ProductCreate":{"properties":{"name":{"type":"string","maxLength":100,"title":"Name"},"price_cents":{"type":"integer","minimum":0.0,"title":"Price Cents"},"stock":{"type":"integer","minimum":0.0,"title":"Stock"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string","maxLength":2083,"minLength":1,"format":"uri"},{"type":"null"}],"title":"Image Url"},"category_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"},"sku":{"type":"string","maxLength":32,"title":"Sku"}},"type":"object","required":["name","price_cents","stock","description","image_url","category_id","sku"],"title":"ProductCreate"},"ProductOut":{"properties":{"name":{"type":"string","maxLength":100,"title":"Name"},"price_cents":{"type":"integer","minimum":0.0,"title":"Price Cents"},"stock":{"type":"integer","minimum":0.0,"title":"Stock"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string","maxLength":2083,"minLength":1,"format":"uri"},{"type":"null"}],"title":"Image Url"},"category_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"},"sku":{"type":"string","title":"Sku"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["name","price_cents","stock","description","image_url","category_id","sku","created_at","updated_at"],"title":"ProductOut"},"ProductUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"price_cents":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Price Cents"},"stock":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Stock"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"image_url":{"anyOf":[{"type":"string","maxLength":2083,"minLength":1,"format":"uri"},{"type":"null"}],"title":"Image Url"},"category_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"}},"type":"object","required":["name","price_cents","stock","description","image_url","category_id"],"title":"ProductUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserCreate":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","minLength":6,"title":"Password"}},"type":"object","required":["username","email","password"],"title":"UserCreate"},"UserOut":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"email":{"type":"string","format":"email","title":"Email"},"user_id":{"type":"integer","title":"User Id"},"is_admin":{"type":"boolean","title":"Is Admin"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["username","email","user_id","is_admin","created_at"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/token"}}}}}}